+++ CHANGELOG.md
@@ -12,6 +12,8 @@ v3.1.0 - TBD
 * Improves the Galileo API: Shapes can now be drawn directly, and Groups now
   have a `transform` property which allows the transformation matrix to be
   modified directly.  The `x`/`y` and `angle` properties have been removed.
+* `Async()` is now called `DispatchScript()` for consistency with other parts of
+  the Sphere API.
 
 
 v3.0.8 - April 17, 2016